requirements.txt
discord.py>=2.5.2
aiosqlite>=0.21.0
pytz>=2025.2